refactor: normalize dialog / modal usage
