+++ src/components/factory/FactoryLayout.svelte
@@ -2,8 +2,8 @@
 	import FactoryPlanner from '@/components/factory/FactoryPlanner.svelte';
 	import FactoryStage from '@/components/factory/FactoryStage.svelte';
 	import OverviewWindow from '@/components/factory/overview/OverviewWindow.svelte';
-	import MachineConfigurationModal from '@/components/shared/MachineConfigurationModal.svelte';
-	import EdgeModal from '@/components/shared/EdgeModal.svelte';
+	import MachineConfigurationDialog from '@/components/shared/MachineConfigurationDialog.svelte';
+	import EdgeDialog from '@/components/shared/EdgeDialog.svelte';
 	import { recalculateEdgeAmounts } from '@/lib/factory/edge';
 	import { type Edge, type MachineConfiguration } from '@/lib/models/factory';
 	import active from '@/stores/active.svelte';
@@ -53,9 +53,9 @@
 	<OverviewWindow />
 {/if}
 
-<MachineConfigurationModal
+<MachineConfigurationDialog
 	bind:dialog={machineConfigDialog}
 	bind:config={editingMachineConfig}
 	onChange={updateMachineConfig}
 />
-<EdgeModal bind:dialog={edgeDialog} bind:edge={selectedEdge} />
+<EdgeDialog bind:dialog={edgeDialog} bind:edge={selectedEdge} />
