+++ src/components/shared/EdgeDialog.svelte
