+++ src/components/shared/MachineConfigurationDialog.svelte
